// src/app/models/test.model.ts
export class TestModel {
    ahorroVoluntario?: number;
    ahorroVoluntarioNecesario?: number;
    captcha: string;
    host: string;
    celular: string;
    correo: string;
    edad?: string;
    estatusAhorro: string;
    estatusContingencia: boolean;
    estatusOcasional: string;
    estatusDeudaRecom: boolean;
    estatusEducacion: string;
    estatusEntretenimiento: string;
    estatusHospitalizacion: boolean;
    estatusPersonales: string;
    estatusResumenRetiro: boolean;
    estatusRetiro: boolean;
    estatusServicios: string;
    estatusTransporte: string;
    estatusVivienda: string;
    evaluacionTest: string;
    gastoAhorro: number;
    gastoOcasional: number;
    gastoEducacion: number;
    gastoEntretenimiento: number;
    gastoPersonales: number;
    gastoServicios: number;
    gastoTransporte: number;
    gastoVivienda: number;
    gastoDeuda: number;
    ingresoMensual: number;
    nombreCompleto: string;
    pensionActual: number;
    pensionDeseada?: number;
    recomendacionContingencia: string;
    recomendacionDeuda: string;
    recomendacionHospitalizacion: string;
    recomendacionRetiro: string;
    respuestaContingencia: string;
    respuestaDeuda: string;
    respuestaHospitalizacion: string;
    respuestaRetiro: string;
    resumenGastosLibresMensuales: string;
    resumenGastosMensuales: string;
    resumenRetiro: string;
    asuntoCorreo: string;
    asesor: any;
}